+++ src/store/data/data.test.js
@@ -7,26 +7,14 @@ import {
   SELECT_CITY,
   FETCH_OFFERS_PENDING,
   FETCH_OFFERS_SUCCESS,
-  SIGN_IN_SUCCESS,
-  GET_USER_SUCCESS,
   FETCH_COMMENTS_PENDING,
   FETCH_COMMENTS_SUCCESS,
   POST_COMMENTS_PENDING,
   POST_COMMENTS_SUCCESS,
-  FETCH_FAVORITE_PENDING,
-  FETCH_FAVORITE_SUCCESS,
+  POST_FAVORITE_SUCCESS,
 } from '../../consts/actionTypes';
 import {ASYNC_STATUSES} from '../../consts/index.js';
 import endpoints from '../../consts/endpoints.js';
-
-describe(`Action creators work correctly`, () => {
-  it(`Action creator for selecting city returns action with payload`, () => {
-    expect(actions.selectCity({fake: true})).toEqual({
-      type: SELECT_CITY,
-      payload: {fake: true},
-    });
-  });
-});
 
 describe(`Reducer works correctly`, () => {
   it(`Reducer without additional parameters should return initial state`, () => {
@@ -38,22 +26,6 @@ describe(`Reducer works correctly`, () => {
       payload: {fake: true},
     })).toEqual(Object.assign({}, initialState, {
       city: {fake: true},
-    }));
-  });
-  it(`Reducer should change user by a given payload`, () => {
-    expect(reducer(undefined, {
-      type: SIGN_IN_SUCCESS,
-      payload: {fake: true},
-    })).toEqual(Object.assign({}, initialState, {
-      user: {fake: true},
-    }));
-  });
-  it(`Reducer should change user by a given payload`, () => {
-    expect(reducer(undefined, {
-      type: GET_USER_SUCCESS,
-      payload: {fake: true},
-    })).toEqual(Object.assign({}, initialState, {
-      user: {fake: true},
     }));
   });
   it(`Reducer should change offers by a given payload`, () => {
@@ -73,7 +45,23 @@ describe(`Reducer works correctly`, () => {
       comments: [{fake: true}],
     }));
   });
+  it(`Reducer should change user by a given payload`, () => {
+    expect(reducer(undefined, {
+      type: POST_FAVORITE_SUCCESS,
+      payload: {offers: [{fake: true}], favorites: [{fake: true}]},
+    })).toEqual(Object.assign({}, initialState, {
+      offers: [{fake: true}],
+    }));
+  });
+});
 
+describe(`Action creators work correctly`, () => {
+  it(`Action creator for selecting city returns action with payload`, () => {
+    expect(actions.selectCity({fake: true})).toEqual({
+      type: SELECT_CITY,
+      payload: {fake: true},
+    });
+  });
   it(`Should make a correct API call to /hotels with get method`, () => {
     const dispatch = jest.fn();
     const api = createApi(dispatch);
@@ -92,45 +80,6 @@ describe(`Reducer works correctly`, () => {
         expect(dispatch).toHaveBeenNthCalledWith(2, {
           type: FETCH_OFFERS_SUCCESS,
           payload: [],
-        });
-      });
-  });
-  it(`Should make a correct API call to /login, with post method`, () => {
-    const dispatch = jest.fn();
-    const api = createApi(dispatch);
-    const apiMock = new MockAdapter(api);
-    const loginUser = actions.authLogin({
-      userEmail: `test@email.com`,
-      userPassword: `qwerty`,
-    });
-
-    apiMock
-      .onPost(endpoints.login)
-      .reply(200, {fake: true});
-
-    return loginUser(dispatch, jest.fn(), api)
-      .then(() => {
-        expect(dispatch).toHaveBeenNthCalledWith(1, {
-          type: SIGN_IN_SUCCESS,
-          payload: {fake: true},
-        });
-      });
-  });
-  it(`Should make a correct API call to /login, with get method`, () => {
-    const dispatch = jest.fn();
-    const api = createApi(dispatch);
-    const apiMock = new MockAdapter(api);
-    const userLoader = actions.getUser();
-
-    apiMock
-      .onGet(endpoints.login)
-      .reply(200, {fake: true});
-
-    return userLoader(dispatch, jest.fn(), api)
-      .then(() => {
-        expect(dispatch).toHaveBeenNthCalledWith(1, {
-          type: GET_USER_SUCCESS,
-          payload: {fake: true},
         });
       });
   });
@@ -176,27 +125,6 @@ describe(`Reducer works correctly`, () => {
         });
         expect(dispatch).toHaveBeenNthCalledWith(2, {
           type: POST_COMMENTS_SUCCESS,
-          payload: [],
-        });
-      });
-  });
-  it(`Should make a correct API call to /favorite with get method`, () => {
-    const dispatch = jest.fn();
-    const api = createApi(dispatch);
-    const apiMock = new MockAdapter(api);
-    const postFavorite = actions.getFavorite();
-
-    apiMock
-      .onGet(endpoints.getFavorite)
-      .reply(200, []);
-
-    return postFavorite(dispatch, jest.fn(), api)
-      .then(() => {
-        expect(dispatch).toHaveBeenNthCalledWith(1, {
-          type: FETCH_FAVORITE_PENDING,
-        });
-        expect(dispatch).toHaveBeenNthCalledWith(2, {
-          type: FETCH_FAVORITE_SUCCESS,
           payload: [],
         });
       });
