+++ src/store/reducer.js
@@ -2,8 +2,10 @@ import {combineReducers} from 'redux';
 
 import rData from './data/data.js';
 import rFilters from './filters/filters.js';
+import rUser from './user/user.js';
 
 export default combineReducers({
   rData,
   rFilters,
+  rUser,
 });
